deduplicate code
